fix(tests): restore the build and dotnet test after the xunit v3 4.0.0 upgrade
The package upgrade left the tree unable to build or run a single test.

Build: xunit.v3 4.0.0 deprecates CollectionBehavior.DisableTestParallelization
as an error, so the solution failed with CS0619. ParallelMode.None is the
documented replacement -- "Disables all parallelism within the test assembly" --
which is the guarantee those plugin tests rely on, since they drive one shared
PluginManager and a live socket per venue.

dotnet test: 4.0.0 drops the VSTest bridge on the .NET 10 SDK, so every test
project refused with "Testing with VSTest target is no longer supported". The
opt-in lives in global.json and nowhere else; a per-project
TestingPlatformDotnetTestSupport property does not work, measured.

Two consequences of that switch worth knowing before relying on them:
coverlet.collector and xunit.runner.visualstudio are VSTest components and are
inert for CLI runs, and --logger/--collect/--settings are no longer accepted
(--filter still is).

Also: wire-number parsing had no tests here at all, which for a helper whose
whole job is a correctness constraint is the gap most likely to be undone by a
future runtime change. Adds the culture tests and the connector scan that keeps
them honest -- the scan finds 10 locale-dependent call sites at b3360f2 and 0
now, so it is a check that can fail.

And in Binance's DecimalConverter, a bare JSON number was read with GetInt32(),
which cannot represent a fraction and throws past Int32's range; the throw comes
back as a JsonException, which Parse catches and turns into a dropped update.
Reads it as a decimal, the way the KuCoin converter already does.

Gemini: RunReconnectAsync carried two consecutive <summary> blocks, the first of
which describes RequestReconnect. Moved onto the method it describes.

Build: dotnet build VisualHFT.sln -c Release -> 0 errors.
Tests: Commons 81/81, MarketResilience 53/53. DataRetriever and TriggerService
have batch-order failures that pass individually, and TriggerService had more of
them before this change than after.

+++ VisualHFT.Plugins/MarketConnectors.Binance/JsonParser.cs
@@ -87,7 +87,12 @@ namespace MarketConnectors.Binance
             }
             else if (reader.TokenType == JsonTokenType.Number)
             {
-                return reader.GetInt32();
+                // GetDecimal, not GetInt32: a bare number token carrying a fraction or a value wider
+                // than Int32 makes GetInt32 throw, which System.Text.Json rethrows as a JsonException
+                // -- caught by Parse above, which then returns null and drops the whole update with
+                // nothing but a Console line to show for it. The KuCoin converter already reads this
+                // token with GetDecimal.
+                return reader.GetDecimal();
             }
 
             return 0;
